# frontend/openapi.json
{
    "openapi": "3.0.1",
    "info": {
        "title": "业务监控系统 API",
        "description": "业务监控系统后端接口文档，包含服务器资产管理、工作流编排、告警配置等功能模块。",
        "contact": {
            "name": "Monitor Team",
            "email": "monitor@example.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "https://www.apache.org/licenses/LICENSE-2.0"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8080",
            "description": "开发环境"
        }
    ],
...
    "paths": {
...
        "/api/server-group": {
            "get": {
                "tags": [
                    "服务器分组"
                ],
                "summary": "获取所有分组",
                "operationId": "listAll",
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiResponseListServerGroup"
                                }
                            }
                        }
                    }
                }
            },
            "put": {
                "tags": [
                    "服务器分组"
                ],
                "summary": "更新分组",
                "operationId": "update_4",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/ServerGroupRequest"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiResponseVoid"
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "服务器分组"
                ],
                "summary": "添加分组",
                "operationId": "add_3",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/ServerGroupRequest"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiResponseVoid"
                                }
                            }
                        }
                    }
                }
            }
        },
...
        "/api/server-monitor-task/server/{serverId}": {
            "get": {
                "tags": [
                    "服务器监控任务"
                ],
                "summary": "按服务器获取任务",
                "operationId": "getByServer",
                "parameters": [
                    {
                        "name": "serverId",
                        "in": "path",
                        "description": "服务器ID",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiResponseListServerMonitorTask"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/server-group/{id}": {
            "get": {
                "tags": [
                    "服务器分组"
                ],
                "summary": "获取分组详情",
                "operationId": "getById_3",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "分组ID",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiResponseServerGroup"
                                }
                            }
                        }
                    }
                }
            },
            "delete": {
                "tags": [
                    "服务器分组"
                ],
                "summary": "删除分组",
                "operationId": "delete_4",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "分组ID",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiResponseVoid"
                                }
                            }
                        }
                    }
                }
            }
        },
...
        "/api/server-asset/group/{groupId}": {
            "get": {
                "tags": [
                    "服务器资产管理"
                ],
                "summary": "按分组查询服务器",
                "description": "根据分组ID查询服务器列表",
                "operationId": "listByGroup",
                "parameters": [
                    {
                        "name": "groupId",
                        "in": "path",
                        "description": "分组ID",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiResponseListServerAsset"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "ApiResponseVoid": {
                "type": "object",
                "properties": {
                    "success": {
                        "type": "boolean",
                        "description": "是否成功"
                    },
                    "code": {
                        "type": "integer",
                        "description": "状态码",
                        "format": "int32"
                    },
                    "message": {
                        "type": "string",
                        "description": "消息"
                    },
                    "data": {
                        "type": "object",
                        "description": "响应数据"
                    }
                },
                "description": "统一API响应"
            },
...
            "ServerGroupRequest": {
                "type": "object",
                "properties": {
                    "id": {
                        "type": "integer",
                        "description": "主键ID，更新时必填",
                        "format": "int64"
                    },
                    "name": {
                        "type": "string",
                        "description": "分组名称"
                    },
                    "parentId": {
                        "type": "integer",
                        "description": "父分组ID（顶级分组为null）",
                        "format": "int64"
                    },
                    "description": {
                        "type": "string",
                        "description": "描述"
                    },
                    "isActive": {
                        "type": "integer",
                        "description": "是否启用（1=启用，0=禁用）",
                        "format": "int32"
                    }
                },
                "description": "服务器分组请求"
            },
...
            "ApiResponseListServerMonitorTask": {
                "type": "object",
                "properties": {
                    "success": {
                        "type": "boolean",
                        "description": "是否成功"
                    },
                    "code": {
                        "type": "integer",
                        "description": "状态码",
                        "format": "int32"
                    },
                    "message": {
                        "type": "string",
                        "description": "消息"
                    },
                    "data": {
                        "type": "array",
                        "description": "响应数据",
                        "items": {
                            "$ref": "#/components/schemas/ServerMonitorTask"
                        }
                    }
                },
                "description": "统一API响应"
            },
            "ServerMonitorTask": {
                "type": "object",
                "properties": {
                    "id": {
                        "type": "integer",
                        "description": "主键ID",
                        "format": "int64"
                    },
                    "name": {
                        "type": "string",
                        "description": "任务名称"
                    },
                    "serverId": {
                        "type": "integer",
                        "description": "关联的服务器ID",
                        "format": "int64"
                    },
                    "templateId": {
                        "type": "integer",
                        "description": "关联的模板ID",
                        "format": "int64"
                    },
                    "collectScript": {
                        "type": "string",
                        "description": "采集脚本（Shell命令）"
                    },
                    "params": {
                        "type": "string",
                        "description": "脚本参数（JSON格式）"
                    },
                    "thresholdRule": {
                        "type": "string",
                        "description": "告警阈值规则（JSON格式）"
                    },
                    "cronExpression": {
                        "type": "string",
                        "description": "Cron定时表达式"
                    },
                    "alarmChannels": {
                        "type": "string",
                        "description": "告警渠道ID列表（逗号分隔）"
                    },
                    "alarmTemplateId": {
                        "type": "integer",
                        "description": "告警模板ID",
                        "format": "int64"
                    },
                    "isActive": {
                        "type": "integer",
                        "description": "是否启用（1=启用，0=禁用）",
                        "format": "int32"
                    },
                    "lastRunTime": {
                        "type": "string",
                        "description": "最后执行时间",
                        "format": "date-time"
                    },
                    "lastRunStatus": {
                        "type": "string",
                        "description": "最后执行状态（SUCCESS/FAILED）"
                    },
                    "lastRunValue": {
                        "type": "string",
                        "description": "最后执行结果值"
                    },
                    "createTime": {
                        "type": "string",
                        "description": "创建时间",
                        "format": "date-time"
                    },
                    "updateTime": {
                        "type": "string",
                        "description": "更新时间",
                        "format": "date-time"
                    },
                    "serverIp": {
                        "type": "string"
                    },
                    "templateName": {
                        "type": "string"
                    },
                    "serverName": {
                        "type": "string"
                    }
                },
                "description": "服务器监控任务"
            },
...
            "ApiResponseListServerGroup": {
                "type": "object",
                "properties": {
                    "success": {
                        "type": "boolean",
                        "description": "是否成功"
                    },
                    "code": {
                        "type": "integer",
                        "description": "状态码",
                        "format": "int32"
                    },
                    "message": {
                        "type": "string",
                        "description": "消息"
                    },
                    "data": {
                        "type": "array",
                        "description": "响应数据",
                        "items": {
                            "$ref": "#/components/schemas/ServerGroup"
                        }
                    }
                },
                "description": "统一API响应"
            },
            "ServerGroup": {
                "type": "object",
                "properties": {
                    "id": {
                        "type": "integer",
                        "description": "主键ID",
                        "format": "int64"
                    },
                    "name": {
                        "type": "string",
                        "description": "分组名称"
                    },
                    "description": {
                        "type": "string",
                        "description": "分组描述"
                    },
                    "parentId": {
                        "type": "integer",
                        "description": "父分组ID（顶级分组为null）",
                        "format": "int64"
                    },
                    "isActive": {
                        "type": "integer",
                        "description": "是否启用（1=启用，0=禁用）",
                        "format": "int32"
                    },
                    "createTime": {
                        "type": "string",
                        "description": "创建时间",
                        "format": "date-time"
                    },
                    "updateTime": {
                        "type": "string",
                        "description": "更新时间",
                        "format": "date-time"
                    }
                },
                "description": "服务器分组"
            },
            "ApiResponseServerGroup": {
                "type": "object",
                "properties": {
                    "success": {
                        "type": "boolean",
                        "description": "是否成功"
                    },
                    "code": {
                        "type": "integer",
                        "description": "状态码",
                        "format": "int32"
                    },
                    "message": {
                        "type": "string",
                        "description": "消息"
                    },
                    "data": {
                        "$ref": "#/components/schemas/ServerGroup"
                    }
                },
                "description": "统一API响应"
            },
            "ApiResponseListServerAsset": {
                "type": "object",
                "properties": {
                    "success": {
                        "type": "boolean",
                        "description": "是否成功"
                    },
                    "code": {
                        "type": "integer",
                        "description": "状态码",
                        "format": "int32"
                    },
                    "message": {
                        "type": "string",
                        "description": "消息"
                    },
                    "data": {
                        "type": "array",
                        "description": "响应数据",
                        "items": {
                            "$ref": "#/components/schemas/ServerAsset"
                        }
                    }
                },
                "description": "统一API响应"
            },
            "ServerAsset": {
                "type": "object",
                "properties": {
                    "id": {
                        "type": "integer",
                        "description": "主键ID",
                        "format": "int64"
                    },
                    "name": {
                        "type": "string",
                        "description": "服务器名称"
                    },
                    "ip": {
                        "type": "string",
                        "description": "服务器IP地址"
                    },
                    "port": {
                        "type": "integer",
                        "description": "SSH端口号",
                        "format": "int32"
                    },
                    "username": {
                        "type": "string",
                        "description": "SSH用户名"
                    },
                    "authType": {
                        "type": "string",
                        "description": "认证类型（PASSWORD/KEY）"
                    },
                    "groupId": {
                        "type": "integer",
                        "description": "所属分组ID",
                        "format": "int64"
                    },
                    "tags": {
                        "type": "string",
                        "description": "标签（逗号分隔）"
                    },
                    "isActive": {
                        "type": "integer",
                        "description": "是否启用（1=启用，0=禁用）",
                        "format": "int32"
                    },
                    "lastCheckTime": {
                        "type": "string",
                        "description": "最后健康检查时间",
                        "format": "date-time"
                    },
                    "lastCheckStatus": {
                        "type": "string",
                        "description": "最后健康检查状态（ONLINE/OFFLINE）"
                    },
                    "createTime": {
                        "type": "string",
                        "description": "创建时间",
                        "format": "date-time"
                    },
                    "updateTime": {
                        "type": "string",
                        "description": "更新时间",
                        "format": "date-time"
                    },
                    "privateKey": {
                        "type": "string"
                    },
                    "password": {
                        "type": "string"
                    }
                },
                "description": "服务器资产"
            },
...
        }
    }
}